requirements.txt
streamlit
pandas
mygene
openpyxl
xlsxwriter
gprofiler-official
plotly